Add the repro for #292 as test
