--- src/Log4NetConcurrencyRepro/Log4NetConcurrencyRepro.csproj
@@ -1,18 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <TargetFrameworks>net472;net10.0-windows</TargetFrameworks>
-    <IsPackable>false</IsPackable>
-    <IsTestProject>true</IsTestProject>
-    <Log4NetVersion Condition="'$(Log4NetVersion)' == ''">3.3.0</Log4NetVersion>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="log4net" Version="$(Log4NetVersion)" />
-    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.10.0" />
-    <PackageReference Include="xunit" Version="2.9.0" />
-    <PackageReference Include="xunit.runner.visualstudio" Version="2.8.1">
-      <PrivateAssets>all</PrivateAssets>
-      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-    </PackageReference>
-  </ItemGroup>
-</Project>
